Fix CI + Devin review: move ParseTrustProxy to config package, fix GoDoc
- Move ParseTrustProxy from server to config package to avoid
  dragging down server package coverage (was 1.4%)
- Fix GoDoc comment ordering: DefaultTrustProxy was mixed into
  deriveMediaProxySecret's doc block
- Config package coverage: 96.4%
