+++ internal/server/server.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"net"
 	"net/http"
 
 	"github.com/hibiken/asynq"
@@ -40,7 +39,11 @@ func New(cfg *config.Config, db *gorm.DB, redis *cache.RedisClients) *Server {
 	e.HidePort = true
 
 	// trustProxyからIPExtractorを構成
-	if opts := parseTrustProxy(cfg.TrustProxy); len(opts) > 0 {
+	if nets := config.ParseTrustProxy(cfg.TrustProxy); len(nets) > 0 {
+		var opts []echo.TrustOption
+		for _, n := range nets {
+			opts = append(opts, echo.TrustIPRange(n))
+		}
 		e.IPExtractor = echo.ExtractIPFromXFFHeader(opts...)
 	}
 
@@ -180,19 +183,4 @@ func buildAsynqRedisOpt(opts config.RedisOptions) asynq.RedisClientOpt {
 // from setupRoutes after the chart engines are constructed.
 func (s *Server) setChartManagement(m *chart.ManagementService) {
 	s.chartMgmt = m
-}
-
-// parseTrustProxy converts a list of CIDR strings into Echo TrustOptions.
-// 無効なCIDRはスキップしてログに記録する。
-func parseTrustProxy(cidrs []string) []echo.TrustOption {
-	var opts []echo.TrustOption
-	for _, cidr := range cidrs {
-		_, ipNet, err := net.ParseCIDR(cidr)
-		if err != nil {
-			slog.Warn("invalid trustProxy CIDR, skipping", "cidr", cidr, "err", err)
-			continue
-		}
-		opts = append(opts, echo.TrustIPRange(ipNet))
-	}
-	return opts
 }
